LayoutTests/imported/w3c/web-platform-tests/web-animations/interfaces/Animatable/getAnimations-iframe-expected.txt



FAIL Calling getAnimations updates layout of parent frame if needed assert_equals: expected 1 but got 0
